Habia diferentes pequeños errores

+++ libft/ft_bzero.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *str, size_t c)
+void	ft_bzero(void *str, int c)
 {
 	int				i;
 	unsigned char	*ptr;
